Removed UI dead code, reorganized some of the DTOs to make it less confusing for me.
