--- Kavita.Models/DTOs/ReadingLists/CBL/CblConflictsDto.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Kavita.Models.DTOs.ReadingLists.CBL;
-
-
-public sealed record CblConflictQuestion
-{
-    public string SeriesName { get; set; }
-    public IList<int> LibrariesIds { get; set; }
-}
