+++ Kavita.Models/DTOs/ReadingLists/CBL/Import/CblReValidateRequestDto.cs
@@ -1,4 +1,4 @@
-namespace Kavita.Models.DTOs.ReadingLists.CBL;
+namespace Kavita.Models.DTOs.ReadingLists.CBL.Import;
 
 /// <summary>
 /// Request body for the re-validate endpoint
