+++ Kavita.Models/DTOs/ReadingLists/CBL/Import/CblSeriesCandidate.cs
@@ -1,0 +1,6 @@
+﻿namespace Kavita.Models.DTOs.ReadingLists.CBL.Import;
+
+/// <summary>
+/// A candidate series for user disambiguation when multiple series match a CBL name
+/// </summary>
+public sealed record CblSeriesCandidate(int SeriesId, int LibraryId, string SeriesName);
