+++ Kavita.Models/DTOs/ReadingLists/CBL/Internal/CblIssueType.cs
@@ -1,4 +1,4 @@
-namespace Kavita.Models.DTOs.ReadingLists.CBL;
+namespace Kavita.Models.DTOs.ReadingLists.CBL.Internal;
 
 /// <summary>
 /// Categorisation of an issue's role within a reading list (V2 only).
