+++ Kavita.Services/ReadingLists/CblImportService.cs
@@ -10,6 +10,8 @@ using Kavita.API.Services.ReadingLists;
 using Kavita.Common.Extensions;
 using Kavita.Models.Builders;
 using Kavita.Models.DTOs.ReadingLists.CBL;
+using Kavita.Models.DTOs.ReadingLists.CBL.Import;
+using Kavita.Models.DTOs.ReadingLists.CBL.Internal;
 using Kavita.Models.Entities;
 using Kavita.Models.Entities.Enums.ReadingList;
 using Kavita.Models.Entities.ReadingLists;
