+++ UI/Web/src/app/_models/reading-list/cbl/cbl-import-summary.ts
@@ -1,10 +1,5 @@
-import { CblBookResult } from "./cbl-book-result";
-import { CblImportResult } from "./cbl-import-result.enum";
-
-export interface CblConflictQuestion {
-    seriesName: string;
-    librariesIds: Array<number>;
-}
+import {CblBookResult} from "./cbl-book-result";
+import {CblImportResult} from "./cbl-import-result.enum";
 
 export interface CblImportSummary {
     cblName: string;
