+++ UI/Web/src/app/_services/reading-list.service.ts
@@ -6,7 +6,6 @@ import {UtilityService} from '../shared/_services/utility.service';
 import {Person, PersonRole} from '../_models/metadata/person';
 import {PaginatedResult} from '../_models/pagination';
 import {ReadingList, ReadingListCast, ReadingListInfo, ReadingListItem} from '../_models/reading-list';
-import {CblImportSummary} from '../_models/reading-list/cbl/cbl-import-summary';
 import {TextResonse} from '../_types/text-response';
 import {ActionItem} from "../_models/actionables/action-item";
 import {Action} from "../_models/actionables/action";
@@ -103,14 +102,6 @@ export class ReadingListService {
 
   nameExists(name: string) {
     return this.httpClient.get<boolean>(this.baseUrl + 'readinglist/name-exists?name=' + name);
-  }
-
-  validateCbl(form: FormData, dryRun: boolean, useComicVineMatching: boolean) {
-    return this.httpClient.post<CblImportSummary>(this.baseUrl + `cbl/validate?dryRun=${dryRun}&useComicVineMatching=${useComicVineMatching}`, form);
-  }
-
-  importCbl(form: FormData, dryRun: boolean, useComicVineMatching: boolean) {
-    return this.httpClient.post<CblImportSummary>(this.baseUrl + `cbl/import?dryRun=${dryRun}&useComicVineMatching=${useComicVineMatching}`, form);
   }
 
   getPeople(readingListId: number, role: PersonRole) {
