--- UI/Web/src/app/reading-list/_components/import-cbl/import-cbl.component.html
@@ -1,183 +1,0 @@
-<ng-container *transloco="let t; prefix: 'import-cbl'">
-
-  <div class="row g-0" style="min-width: 8.4375rem;">
-    <app-step-tracker [steps]="steps" [currentStep]="currentStepIndex" />
-  </div>
-
-  <div>
-    @switch (currentStepIndex) {
-      @case (Step.Import) {
-        <div class="row g-0">
-          <p>{{t('import-description')}}</p>
-          <p [innerHTML]="t('cbl-repo') | safeHtml"></p>
-
-          <form [formGroup]="uploadForm" enctype="multipart/form-data">
-            <file-upload formControlName="files" />
-          </form>
-        </div>
-      }
-      @case (Step.Validate) {
-        <p>{{t('validate-description')}}</p>
-        <div class="row g-0">
-          <div ngbAccordion #accordion="ngbAccordion">
-            @for(fileToProcess of filesToProcess; track fileToProcess.fileName) {
-              @if (fileToProcess.validateSummary; as summary) {
-                <div ngbAccordionItem>
-                  <h5 ngbAccordionHeader>
-                    <button ngbAccordionButton class="accordion-button">
-                      <ng-container [ngTemplateOutlet]="heading" [ngTemplateOutletContext]="{ summary: summary, filename: fileToProcess.fileName }" />
-                      <i class="fas fa-chevron-up" aria-hidden="true"></i>
-                    </button>
-                  </h5>
-                  <div ngbAccordionCollapse>
-                    <div ngbAccordionBody>
-                      <ng-container [ngTemplateOutlet]="validationList" [ngTemplateOutletContext]="{ summary: summary }" />
-                    </div>
-                  </div>
-                </div>
-              }
-            }
-          </div>
-        </div>
-      }
-      @case (Step.DryRun) {
-        <div class="row g-0">
-          <p>{{t('dry-run-description')}}</p>
-
-          <div ngbAccordion #a="ngbAccordion">
-            @for(fileToProcess of filesToProcess; track fileToProcess.fileName) {
-              @if (fileToProcess.dryRunSummary; as summary) {
-                <div ngbAccordionItem>
-                  <h5 ngbAccordionHeader>
-                    <button ngbAccordionButton class="accordion-button">
-                      <ng-container [ngTemplateOutlet]="heading" [ngTemplateOutletContext]="{ summary: summary, filename: fileToProcess.fileName }" />
-                      <i class="fas fa-chevron-up" aria-hidden="true"></i>
-                    </button>
-                  </h5>
-                  <div ngbAccordionCollapse>
-                    <div ngbAccordionBody>
-                      <ng-container [ngTemplateOutlet]="resultsList" [ngTemplateOutletContext]="{ summary: summary }" />
-                    </div>
-                  </div>
-                </div>
-              }
-            }
-          </div>
-        </div>
-      }
-      @case (Step.Finalize) {
-        <div class="row g-0">
-          <div ngbAccordion #a="ngbAccordion">
-            @for(fileToProcess of filesToProcess; track fileToProcess.fileName) {
-              @if (fileToProcess.finalizeSummary; as summary) {
-                <div ngbAccordionItem>
-                  <h5 ngbAccordionHeader>
-                    <button ngbAccordionButton class="accordion-button">
-                      <ng-container [ngTemplateOutlet]="heading" [ngTemplateOutletContext]="{ summary: summary, filename: fileToProcess.fileName }" />
-                      <i class="fas fa-chevron-up" aria-hidden="true"></i>
-                    </button>
-                  </h5>
-                  <div ngbAccordionCollapse>
-                    <div ngbAccordionBody>
-                      <ng-container [ngTemplateOutlet]="resultsList" [ngTemplateOutletContext]="{ summary: summary }" />
-                    </div>
-                  </div>
-                </div>
-              }
-            }
-          </div>
-        </div>
-      }
-    }
-  </div>
-
-  <ng-template #validationList let-summary="summary">
-    @if (summary.results.length > 0) {
-      <div class="justify-content-center col">
-        <div class="d-flex align-items-center">
-          <div class="flex-shrink-0">
-            <i class="fa-solid fa-triangle-exclamation" style="font-size: 1.5rem" aria-hidden="true"></i>
-          </div>
-          <div class="flex-grow-1 ms-3">
-            {{t('validate-warning')}}
-          </div>
-        </div>
-      </div>
-      <ol class="list-group list-group-numbered list-group-flush" >
-        @for(result of summary.results; track result) {
-          <li class="list-group-item no-hover"
-              [innerHTML]="result | cblConflictReason | safeHtml">
-          </li>
-        }
-
-
-      </ol>
-    }
-    @else {
-      <div class="justify-content-center col">
-        <div class="d-flex align-items-center">
-          <div class="flex-shrink-0">
-            <i class="fa-solid fa-circle-check" style="font-size: 1.5rem" aria-hidden="true"></i>
-          </div>
-          <div class="flex-grow-1 ms-3">
-            {{t('validate-no-issue-description')}}
-          </div>
-        </div>
-      </div>
-    }
-  </ng-template>
-
-  <ng-template #resultsList let-summary="summary">
-    <ul class="list-group list-group-flush">
-      @for(result of summary.results; track result.order) {
-        <li class="list-group-item no-hover"
-            innerHTML="{{result.order + 1}}. {{result | cblConflictReason | safeHtml}}"></li>
-      }
-    </ul>
-  </ng-template>
-
-  <ng-template #heading let-filename="filename" let-summary="summary">
-    @switch (summary.success) {
-      @case (CblImportResult.Success) {
-        <span class="badge heading-badge bg-primary me-1">{{summary.success | cblImportResult}}</span>
-      }
-      @case (CblImportResult.Fail) {
-        <span class="badge heading-badge bg-danger me-1">{{summary.success | cblImportResult}}</span>
-      }
-      @case (CblImportResult.Partial) {
-        <span class="badge heading-badge bg-warning me-1">{{summary.success | cblImportResult}}</span>
-      }
-    }
-    <span>{{filename}}
-      @if(summary.cblName) {
-        <span>: ({{summary.cblName}})</span>
-      }
-    </span>
-  </ng-template>
-
-<div class="modal-footer mt-3">
-  <form [formGroup]="cblSettingsForm" class="row align-items-center">
-    <div class="col-auto">
-      <div class="form-check form-switch">
-        <input type="checkbox" id="settings-comicvine-mode" role="switch" formControlName="comicVineMatching" class="form-check-input"
-               aria-labelledby="auto-close-label" switch>
-        <label class="form-check-label" for="settings-comicvine-mode">{{t('comicvine-parsing-label')}}</label>
-      </div>
-    </div>
-  </form>
-  <!-- Spacer -->
-  <div class="col" aria-hidden="true"></div>
-  <div class="col-auto ms-1">
-    <a class="btn btn-icon" [href]="WikiLink.ReadingListCBL" target="_blank" rel="noopener noreferrer">{{t('help-label')}}</a>
-  </div>
-  <div class="col-auto ms-1">
-    <button type="button" class="btn btn-primary" (click)="prevStep()" [disabled]="!canMoveToPrevStep()">{{t('prev')}}</button>
-  </div>
-  <div class="col-auto ms-1">
-    <button type="button" class="btn btn-primary" (click)="nextStep()" [disabled]="!canMoveToNextStep()">{{t(NextButtonLabel)}}</button>
-  </div>
-</div>
-
-
-
-</ng-container>
