--- UI/Web/src/app/reading-list/_components/import-cbl/import-cbl.component.scss
@@ -1,46 +1,0 @@
-.file-input {
-  display: none;
-}
-
-.heading-badge  {
-  color: var(--bs-badge-color);
-}
-
-::ng-deep .file-info {
-  width: 83%;
-  float: left;
-}
-
-::ng-deep .file-buttons {
-  float: right;
-}
-
-file-upload {
-  background: none;
-  height: auto;
-}
-
-::ng-deep .upload-input {
-  color: var(--input-text-color) !important;
-}
-
-::ng-deep file-upload-list-item {
-  color: var(--input-text-color) !important;
-}
-
-::ng-deep .remove-btn {
-  background: #C0392B;
-  border-radius: 0.1875rem;
-  color: var(--input-text-color) !important;
-  font-weight: bold;
-  padding: 0.1875rem 0.3125rem;
-}
-
-::ng-deep .reading-list-success--item {
-  color: var(--primary-color);
-}
-
-::ng-deep .reading-list-fail--item {
-  color: var(--error-color);
-}
-
