+++ UI/Web/src/app/sidenav/preference-nav/preference-nav.component.ts
@@ -69,7 +69,6 @@ export enum SettingsTabId {
   ScrobblingHolds = 'scrobble-holds',
   Customize = 'customize',
   CBLImport = 'cbl-import',
-  CBLImportV2 = 'cbl-import-v2',
 }
 
 export enum SettingSectionId {
@@ -269,7 +268,6 @@ export class PreferenceNavComponent implements AfterViewInit {
         children: [
           new SideNavItem(SettingsTabId.MappingsImport, [Role.Admin]),
           new SideNavItem(SettingsTabId.CBLImport, [], undefined, [Role.ReadOnly]),
-          new SideNavItem(SettingsTabId.CBLImportV2, [], undefined, [Role.ReadOnly]),
           SideNavItem.kPlusOnly(SettingsTabId.MALStackImport),
         ]
       },
